sprint-hub/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sprint-hub"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "textual>=8.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
]

[project.scripts]
sprint-hub     = "sprint_hub.cli:hub"
sprint-init    = "sprint_hub.cli:init"
sprint-add     = "sprint_hub.cli:add"
sprint-capture = "sprint_hub.cli:capture"
sprint-remove  = "sprint_hub.cli:remove"
sprint-relabel = "sprint_hub.cli:relabel"
sprint-edit    = "sprint_hub.cli:edit_entry"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
